pyproject.toml
[tool.poetry]
name = "course-work-site-ads"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "course_work_site_ads"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
drf-spectacular = "^0.26.0"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^3.14.0"
djangorestframework = "^3.14.0"
djoser = "^2.1.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
